app/requirements.txt
fastapi
uvicorn[standard]
pandas
matplotlib
pydantic
pytest
boto3
fastapi-cors
python-multipart
sqlalchemy
databases
psycopg2-binary
pymysql
pydantic-settings
python-dotenv
passlib==1.7.4
alembic

